requirements.txt
Django==3.1.4
djangorestframework==3.12.2
django-cors-headers==3.6.0
Pillow==3.4.0
psycopg2==2.8.4
python-dotenv==0.12.0
pytz==2019.3
requests==2.24.0